Implement build-socket only no more separate watching
