--- cli/plasmo/src/features/extension-devtools/build-watcher.ts
@@ -1,49 +1,0 @@
-import { subscribe } from "@parcel/watcher"
-import { WebSocket, WebSocketServer } from "ws"
-
-import { PARCEL_WATCHER_BACKEND } from "@plasmo/constants/misc"
-
-import type { PlasmoManifest } from "~features/manifest-factory/base"
-
-const payload = JSON.stringify({ type: "build_ready" })
-
-const DEBOUNCE_TIME = 740
-
-let buildDebounceTimer: NodeJS.Timeout
-
-let hasPrematureBuild = false
-
-export const createBuildWatcher = async (
-  plasmoManifest: PlasmoManifest,
-  hmrPort: number
-) => {
-  const wss = new WebSocketServer({ port: hmrPort + 1 })
-  return subscribe(
-    plasmoManifest.commonPath.buildDirectory,
-    async (err) => {
-      if (err) {
-        throw err
-      }
-
-      if (!hasPrematureBuild) {
-        hasPrematureBuild = true
-        return
-      }
-
-      clearTimeout(buildDebounceTimer)
-      buildDebounceTimer = setTimeout(() => {
-        clearTimeout(buildDebounceTimer)
-
-        for (const client of wss.clients) {
-          if (client.readyState === WebSocket.OPEN) {
-            client.send(payload)
-          }
-        }
-        hasPrematureBuild = false
-      }, DEBOUNCE_TIME)
-    },
-    {
-      backend: PARCEL_WATCHER_BACKEND
-    }
-  )
-}
